+++ R/age_generator.koch6.R
@@ -5,6 +5,7 @@
 #' of treatment date to generate age variable in years. 
 #' @param x data frame containing Koch 6 admission variables 
 #' @param categorise logical - generate additional factor age variable
+#' @param paediatric logical - generate additional paediatric age factor variable
 #' @param rm_orig remove original variables - TRUE or FALSE
 #' @author Jay Achar 
 #' @seealso \code{\link{tbgeneratr}}
@@ -14,6 +15,7 @@
 
 age_generator.koch6 <- function(x, 
                                 categorise = FALSE,
+                                paediatric = FALSE,
                                 rm_orig = TRUE) {
   
   # check variables are present
@@ -49,6 +51,20 @@ age_generator.koch6 <- function(x,
                         "15 - <= 45y",
                         ">45y"))
       
+  }
+  
+  # generate additional paediatric factor variable
+  if (paediatric) {
+    
+    x$age_paeds <- cut(x$age_years,
+                       breaks = c(0, 2, 5, 12, 18),
+                       labels = c(1:4)) %>% 
+      factor(levels = c(1:4),
+             labels = c("<= 2y",
+                        "2 - <= 5y",
+                        "5 - <= 12y",
+                        "12 - <= 18y"))
+    
   }
   
   # remove original variables