+++ data-raw/testdata_age_generator.R
@@ -3,11 +3,13 @@
 age_gen_koch6 <- data.frame(dateofbirth = lubridate::dmy("1/1/2015",
                                               "5/3/2012",
                                               "7/6/1967",
-                                              "1/1/1950"),
+                                              "1/1/1950",
+                                              "1/1/1990"),
                  Starttre = lubridate::dmy("3/1/2016",
                                            "7/3/2018",
                                            NA_character_,
-                                           "1/1/1995"),
+                                           "1/1/1995",
+                                           "1/1/1992"),
                  stringsAsFactors = FALSE)
 class(age_gen_koch6) <- c(class(age_gen_koch6), "koch6")
 saveRDS(age_gen_koch6, file = "inst/testdata/age_gen_koch6.rds")
@@ -29,11 +31,13 @@ saveRDS(age_gen_koch6_neg, file = "inst/testdata/age_gen_koch6_neg.rds")
 age_gen_epi <- data.frame(BIRTDATE = lubridate::dmy("1/1/2015",
                                               "5/3/2012",
                                               "7/6/1967",
-                                              "1/1/1950"),
+                                              "1/1/1950",
+                                              "1/1/1990"),
                  STARTTRE = lubridate::dmy("3/1/2016",
                                            "7/3/2018",
                                            NA_character_,
-                                           "1/1/1995"),
+                                           "1/1/1995",
+                                           "1/1/1992"),
                  stringsAsFactors = FALSE)
 class(age_gen_epi) <- c(class(age_gen_epi), "epiinfo")
 saveRDS(age_gen_epi, file = "inst/testdata/age_gen_epi.rds")