+++ NEWS.md
@@ -2,6 +2,7 @@
 #### 25 Feb 2019
 ## Minor update
 * `age_generator()`: add `categorise` argument to generate additional factor variable for age. 
+* `age_generator()`: add `paediatric` argument to generate additional paediatric age variable.
 
 ---